+++ app/Http/Controllers/TrainingSessionController.php
@@ -25,6 +25,14 @@ class TrainingSessionController extends Controller
      */
     public function store(StoreUpdateTrainingSessionRequest $request) {
         
+        $trainingSessions = TrainingSession::all();
+
+        foreach ($trainingSessions as $trainingSession) {
+            if (($trainingSession->start_at <= $request->start_at && $trainingSession->finish_at >= $request->start_at) || ($trainingSession->start_at <= $request->finish_at && $trainingSession->finish_at >= $request->finish_at)) {
+                return redirect()->back()->with('error', 'Training session time overlaps with another session.');
+            }
+        }
+
         TrainingSession::create($request->all());
 
         return redirect()->route('tables.training_sessions');
@@ -35,6 +43,10 @@ class TrainingSessionController extends Controller
      * 
      */
     public function update(StoreUpdateTrainingSessionRequest $request, TrainingSession $trainingSession) {
+
+        if($trainingSession->users()->count()) {
+            return redirect()->back()->with('error', 'Can\'t update a training session that has users.');
+        }
 
         $trainingSession->update($request->all());
 
@@ -47,8 +59,12 @@ class TrainingSessionController extends Controller
      */
     public function destroy(TrainingSession $trainingSession) {
 
+        if($trainingSession->users()->count()) {
+            return redirect()->back()->with('error', 'Can\'t delete a training session that has users.');
+        }
+
         $trainingSession->delete();
-        
+
         return redirect()->route('tables.training_sessions');
     }
 }
